feature: make Wi-Fi calling carrier name configurable
