+++ app/src/main/java/dev/bluehouse/enablevolte/pages/Home.kt
@@ -21,16 +21,16 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.Dp
 import androidx.core.content.ContextCompat.startActivity
 import androidx.navigation.NavController
-import dev.bluehouse.enablevolte.BooleanPropertyView
 import dev.bluehouse.enablevolte.BuildConfig
 import dev.bluehouse.enablevolte.CarrierModer
-import dev.bluehouse.enablevolte.ClickablePropertyView
-import dev.bluehouse.enablevolte.HeaderText
 import dev.bluehouse.enablevolte.R
 import dev.bluehouse.enablevolte.ShizukuStatus
-import dev.bluehouse.enablevolte.StringPropertyView
 import dev.bluehouse.enablevolte.SubscriptionModer
 import dev.bluehouse.enablevolte.checkShizukuPermission
+import dev.bluehouse.enablevolte.components.BooleanPropertyView
+import dev.bluehouse.enablevolte.components.ClickablePropertyView
+import dev.bluehouse.enablevolte.components.HeaderText
+import dev.bluehouse.enablevolte.components.StringPropertyView
 import dev.bluehouse.enablevolte.getLatestAppVersion
 import dev.bluehouse.enablevolte.uniqueName
 import net.swiftzer.semver.SemVer
